requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
av==14.4.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
Brotli==1.1.0
build==1.2.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.9.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.0
cssselect2==0.8.0
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
deepl==1.18.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
durationpy==0.10
EbookLib==0.18
effdet==0.4.1
elastic-transport==8.15.0
elasticsearch==8.15.0
emoji==2.12.1
environs==9.5.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.112.2
fastavro==1.9.5
faster-whisper==1.1.1
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.15.4
filetype==1.2.0
FlashRank==0.2.9
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==1.34.1
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google_search_results==2.4.2
googleapis-common-protos==1.65.0
grandalf==0.8
greenlet==3.0.3
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.66.1
grpcio-status==1.48.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.23.5
humanfriendly==10.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
iniconfig==2.1.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-pinecone==0.2.6
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.42
llvmlite==0.44.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.40.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.79.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.1.1
posthog==4.0.1
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pulsar-client==3.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.0
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
Send2Trash==1.8.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.38.6
sympy==1.14.0
syrupy==4.9.1
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
types-python-dateutil==2.9.0.20250516
types-requests==2.32.0.20250515
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
weasyprint==65.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zopfli==0.2.3.post1
zstandard==0.23.0
